all change are ready
