--- manage_observations.py
@@ -1,8 +1,0 @@
-from files.observations import*
-import numpy as np
-import pandas as pd
-from datetime import datetime
-
-
-observations = Observations("observaciones.csv")
-print(observations.PrepareData())
